Create Construction-Use-Hydroxypropyl-Starch-Ether-Market.nd
The global Construction Use Hydroxypropyl Starch Ether market size was estimated at USD 118.50 million in 2023 and is projected to reach USD 191.86 million by 2032, with a CAGR of 5.50% during the forecast period. The North America Construction Use Hydroxypropyl Starch Ether market size was estimated at USD 33.86 million in 2023, with a CAGR of 4.71% from 2025 to 2032.