Create Hydrolyzed-Sodium-Hyaluronate-Market-Global-Outlook-and-Forecast-2025-2032.md
Hydrolyzed sodium hyaluronate is the salt form of hyaluronic acid that has undergone the process of hydrolysis. This transformation involves the addition of water or enzymes to alter the ingredient's form.

The global market size for Hydrolyzed Sodium Hyaluronate was estimated at USD 8652 million in 2023, with a projected growth to reach USD 16727.34 million by 2032, representing a CAGR of 7.60%. The North America market size was estimated at USD 2557.76 million in 2023, with a CAGR of 6.51% during the forecast period.